pyproject.toml
[project]
name = "russian-history-project-1"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "python-docx (>=1.1.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "russian_history_project_1", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
